requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
av==16.0.1
    # via faster-whisper
certifi==2024.8.30
    # via requests
cffi==2.0.0
    # via soundfile
ctranslate2==4.6.1
    # via faster-whisper
charset-normalizer==3.4.0
    # via requests
coloredlogs==15.0.1
    # via
    #   ctranslate2
    #   onnxruntime
faster-whisper==1.2.1
    # via -r requirements.in
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
flatbuffers==25.9.23
    # via onnxruntime
fsspec==2024.10.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.27.1
    # via faster-whisper
humanfriendly==10.0
    # via
    #   coloredlogs
    #   ctranslate2
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
markupsafe==3.0.2
    # via jinja2
more-itertools==10.5.0
    # via openai-whisper
mouseinfo==0.1.3
    # via pyautogui
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==1.26.3
    # via
    #   -r requirements.in
    #   numba
    #   onnxruntime
    #   openai-whisper
    #   soundfile
onnxruntime==1.23.2
    # via
    #   faster-whisper
    #   openwakeword
    #   piper-tts
openai-whisper==20231117
    # via -r requirements.in
openwakeword==0.6.0
    # via -r requirements.in (TICKET-P3-02: Wake word detection)
packaging==25.0
    # via onnxruntime
resemblyzer==0.1.4
    # via -r requirements.in (TICKET-STT-002: Voice fingerprinting)
pillow==10.2.0
    # via
    #   -r requirements.in
    #   pyscreeze
piper-tts==1.3.0
    # via -r requirements.in
platformdirs==4.5.0
    # via -r requirements.in
protobuf==6.33.1
    # via onnxruntime
psutil==5.9.8
    # via -r requirements.in
pyaudio==0.2.14
    # via -r requirements.in
pyautogui==0.9.54
    # via -r requirements.in
pycparser==2.23
    # via cffi
pygetwindow==0.0.9
    # via pyautogui
pymsgbox==1.0.9
    # via pyautogui
pynput==1.7.6
    # via -r requirements.in
pyobjc-core==10.1
    # via
    #   pyautogui
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-cocoa
    #   pyobjc-framework-coretext
    #   pyobjc-framework-quartz
pyobjc-framework-applicationservices==10.1
    # via
    #   -r requirements.in
    #   pynput
pyobjc-framework-cocoa==10.1
    # via
    #   -r requirements.in
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-coretext
    #   pyobjc-framework-quartz
pyobjc-framework-coretext==10.1
    # via pyobjc-framework-applicationservices
pyobjc-framework-quartz==10.1
    # via
    #   pyautogui
    #   pynput
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-coretext
pyperclip==1.8.2
    # via
    #   -r requirements.in
    #   mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
pyside6==6.10.0
    # via -r requirements.in
pyside6-addons==6.10.0
    # via pyside6
pyside6-essentials==6.10.0
    # via
    #   pyside6
    #   pyside6-addons
python-dotenv==1.0.1
    # via -r requirements.in
pytweening==1.2.0
    # via pyautogui
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via tiktoken
rubicon-objc==0.5.2
    # via mouseinfo
shiboken6==6.10.0
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
six==1.16.0
    # via pynput
soundfile==0.13.1
    # via -r requirements.in
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tiktoken==0.8.0
    # via openai-whisper
torch==2.9.1
    # via openai-whisper
tokenizers==0.19.1
    # via faster-whisper
tqdm==4.67.0
    # via
    #   huggingface-hub
    #   openai-whisper
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   torch
urllib3==2.2.3
    # via requests
watchdog==4.0.1
    # via -r requirements.in
webrtcvad==2.0.10
    # via -r requirements.in
sentry-sdk>=2.0.0
# TICKET-APP-002: Reader Mode - Web Content Extraction
trafilatura==1.12.2
    # via -r requirements.in
courlan>=1.2.0
    # via trafilatura
htmldate>=1.8.1
    # via trafilatura
justext>=3.0.1
    # via trafilatura
lxml>=5.2.2
    # via trafilatura
babel>=2.16.0
    # via courlan
tld>=0.13
    # via courlan
dateparser>=1.1.2
    # via htmldate
python-dateutil>=2.9.0
    # via htmldate

# TICKET-MEM-001: Semantic Memory (RAG Local)
chromadb==0.4.24
    # via -r requirements.in
sentence-transformers==3.3.1
    # via -r requirements.in

# TICKET-FEAT-002: Task Scheduling & Delayed Actions  
schedule==1.2.0
    # via -r requirements.in
cachetools==5.3.3
    # via -r requirements.in
cryptography==42.0.5
    # via -r requirements.in
